+++ Labyrinth/Assets/Scenes/Nav mesh test.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8463a99c12df82148b48b2b05a42e751
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
