+++ Labyrinth/Assets/Scenes/Nav mesh test.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b4db14c5acd6d4e43909f6d96c6b58b1
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
