+++ Labyrinth/Assets/Scripts/Algorithms/Minotaur/MinosHealth.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b8b3e4c832eaf0f4ca1bb7f4dafc6ca0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
