+++ Labyrinth/Assets/Scripts/Algorithms/Minotaur/MinosSenses.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3e7ed64fe6f9bba4c9ddeb5b1da32ccb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
